+++ src/flows/process_candidates_data.py
@@ -2,7 +2,8 @@ import pandas as pd
 from prefect import flow, task, get_run_logger
 import snscrape.modules.twitter as sntwitter
 from ..common import load_raw_dataset, save_dataset
-from ..preprocess import drop_duplicated_candidates, get_twitter_username
+from ..filters import drop_duplicated_candidates
+from ..feature_extraction import get_twitter_username
 
 
 @task
